+++ test/redox-engine_tests/exceptions_test.rb
@@ -2,20 +2,20 @@ require 'test_helper'
 
 class ExceptionsTest < Minitest::Test
   def test_redox_key_error
-    Redox.configure do |r|
+    RedoxEngine.configure do |r|
       r.api_key = nil
       r.secret = nil
     end
-    assert_raises(Redox::APIKeyError) { redox }
+    assert_raises(RedoxEngine::APIKeyError) { redox }
   end
 
   def test_redox_token_error
-    Redox.configure do |r|
+    RedoxEngine.configure do |r|
       r.api_key = 'bad'
       r.secret = 'keys'
     end
     VCR.use_cassette('client/new/token_invalid_key') do
-      assert_raises(Redox::TokenError) { redox }
+      assert_raises(RedoxEngine::TokenError) { redox }
     end
   end
 end