+++ test/redox-engine_tests/visit_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 class VisitTest < Minitest::Test
   def test_attributes_map_from_response
-    v = Redox::Visit.new(visit)
+    v = RedoxEngine::Visit.new(visit)
     assert_equal([
                    {
                      id: 108