+++ test/test_helper.rb
@@ -1,5 +1,5 @@
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
-require 'redox'
+require 'redox-engine'
 
 require 'minitest/autorun'
 require 'minitest/assertions'