+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.1.3] - 2018-04-04
-- Change Redox::Redox to Redox::Client
+- Change RedoxEngine::RedoxEngine to RedoxEngine::Client
 - Refactor Module to accept configuration
 - Refactor Client to pass Rubocop linting
 - Refactor Client to accept stored token/refresh token